+++ core/tauri/src/ipc/authority.rs
@@ -19,6 +19,9 @@ use tauri_utils::acl::{
   resolved::{Resolved, ResolvedCommand, ResolvedScope, ScopeKey},
   ExecutionContext, Scopes,
 };
+use tauri_utils::platform::Target;
+
+use url::Url;
 
 use crate::{ipc::InvokeError, sealed::ManagerBase, Runtime};
 use crate::{AppHandle, Manager};
@@ -40,7 +43,7 @@ pub enum Origin {
   /// Remote origin.
   Remote {
     /// Remote URL.
-    url: String,
+    url: Url,
   },
 }
 
@@ -58,7 +61,7 @@ impl Origin {
     match (self, context) {
       (Self::Local, ExecutionContext::Local) => true,
       (Self::Remote { url }, ExecutionContext::Remote { url: url_pattern }) => {
-        url_pattern.matches(url)
+        url_pattern.test(url)
       }
       _ => false,
     }
@@ -91,7 +94,7 @@ impl CapabilityBuilder {
       windows: Vec::new(),
       webviews: Vec::new(),
       permissions: Vec::new(),
-      platforms: Vec::new(),
+      platforms: None,
     })
   }
 
@@ -189,6 +192,30 @@ impl CapabilityBuilder {
       .0
       .permissions
       .push(PermissionEntry::ExtendedPermission { identifier, scope });
+    self
+  }
+
+  /// Adds a target platform for this capability.
+  ///
+  /// By default all platforms are applied.
+  pub fn platform(mut self, platform: Target) -> Self {
+    self
+      .0
+      .platforms
+      .get_or_insert_with(Default::default)
+      .push(platform);
+    self
+  }
+
+  /// Adds target platforms for this capability.
+  ///
+  /// By default all platforms are applied.
+  pub fn platforms(mut self, platforms: impl IntoIterator<Item = Target>) -> Self {
+    self
+      .0
+      .platforms
+      .get_or_insert_with(Default::default)
+      .extend(platforms);
     self
   }
 }
@@ -499,7 +526,7 @@ impl RuntimeAuthority {
   }
 }
 
-/// List of allowed and denied objects that match either the command-specific or plugin global scope criterias.
+/// List of allowed and denied objects that match either the command-specific or plugin global scope criteria.
 #[derive(Debug)]
 pub struct ScopeValue<T: ScopeObject> {
   allow: Arc<Vec<Arc<T>>>,
@@ -816,44 +843,7 @@ mod tests {
     let resolved_cmd = vec![ResolvedCommand {
       windows: vec![Pattern::new(window).unwrap()],
       context: ExecutionContext::Remote {
-        url: Pattern::new(url).unwrap(),
-      },
-      ..Default::default()
-    }];
-    let allowed_commands = [(command.to_string(), resolved_cmd.clone())]
-      .into_iter()
-      .collect();
-
-    let authority = RuntimeAuthority::new(
-      Default::default(),
-      Resolved {
-        allowed_commands,
-        ..Default::default()
-      },
-    );
-
-    assert_eq!(
-      authority.resolve_access(
-        command,
-        window,
-        webview,
-        &Origin::Remote { url: url.into() }
-      ),
-      Some(resolved_cmd)
-    );
-  }
-
-  #[test]
-  fn remote_domain_glob_pattern_matches() {
-    let url = "http://tauri.*";
-    let command = "my-command";
-    let window = "main";
-    let webview = "main";
-
-    let resolved_cmd = vec![ResolvedCommand {
-      windows: vec![Pattern::new(window).unwrap()],
-      context: ExecutionContext::Remote {
-        url: Pattern::new(url).unwrap(),
+        url: url.parse().unwrap(),
       },
       ..Default::default()
     }];
@@ -875,7 +865,46 @@ mod tests {
         window,
         webview,
         &Origin::Remote {
-          url: url.replace('*', "studio")
+          url: url.parse().unwrap()
+        }
+      ),
+      Some(resolved_cmd)
+    );
+  }
+
+  #[test]
+  fn remote_domain_glob_pattern_matches() {
+    let url = "http://tauri.*";
+    let command = "my-command";
+    let window = "main";
+    let webview = "main";
+
+    let resolved_cmd = vec![ResolvedCommand {
+      windows: vec![Pattern::new(window).unwrap()],
+      context: ExecutionContext::Remote {
+        url: url.parse().unwrap(),
+      },
+      ..Default::default()
+    }];
+    let allowed_commands = [(command.to_string(), resolved_cmd.clone())]
+      .into_iter()
+      .collect();
+
+    let authority = RuntimeAuthority::new(
+      Default::default(),
+      Resolved {
+        allowed_commands,
+        ..Default::default()
+      },
+    );
+
+    assert_eq!(
+      authority.resolve_access(
+        command,
+        window,
+        webview,
+        &Origin::Remote {
+          url: url.replace('*', "studio").parse().unwrap()
         }
       ),
       Some(resolved_cmd)
@@ -908,7 +937,7 @@ mod tests {
         window,
         webview,
         &Origin::Remote {
-          url: "https://tauri.app".into()
+          url: "https://tauri.app".parse().unwrap()
         }
       )
       .is_none());