+++ core/tauri/src/menu/icon.rs
@@ -8,7 +8,7 @@ use super::run_item_main_thread;
 use super::{IconMenuItem, NativeIcon};
 use crate::menu::IconMenuItemInner;
 use crate::run_main_thread;
-use crate::{menu::MenuId, AppHandle, Image, Manager, Runtime};
+use crate::{image::Image, menu::MenuId, AppHandle, Manager, Runtime};
 
 impl<R: Runtime> IconMenuItem<R> {
   /// Create a new menu item.