+++ core/tauri/src/window/mod.rs
@@ -33,12 +33,13 @@ use crate::{
 };
 #[cfg(desktop)]
 use crate::{
+  image::Image,
   menu::{ContextMenu, Menu, MenuId},
   runtime::{
     window::dpi::{Position, Size},
     UserAttentionType,
   },
-  CursorIcon, Image,
+  CursorIcon,
 };
 
 use serde::Serialize;
@@ -1573,7 +1574,7 @@ impl<R: Runtime> Window<R> {
       .map_err(Into::into)
   }
 
-  /// Determines if this window's native minize button should be enabled.
+  /// Determines if this window's native minimize button should be enabled.
   ///
   /// ## Platform-specific
   ///